requirements.txt
aiohttp==3.8.1
async-timeout==4.0.2
attrs==21.4.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
chardet==4.0.0
discord-py-interactions==4.2.0
distlib==0.3.4
filelock==3.6.0
idna==3.3
multidict==6.0.2
pipenv==2022.5.2
platformdirs==2.5.2
psycopg2==2.9.3
pyodbc==4.0.32
six==1.16.0
typing_extensions==4.2.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
yarl==1.7.2
emoji==1.7.0
